requirements.txt
google-adk
python-dotenv 
gspread 
google-auth 
google-cloud-aiplatform
